Move files into a lib/ directory
