+++ lib/reports/_baseReport.js
