+++ lib/reports/badge.js
