+++ lib/tests/jshint.js
