+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "bin": "index.js",
   "scripts": {
-    "test": "mocha --recursive --require spec/setup.js spec"
+    "test": "mocha --recursive --require test/setup.js test/spec"
   },
   "author": "",
   "license": "ISC",
